Move Problematic trait to common to improve module coupling
