+++ src/app/authorised_agents/pages/view.rs
@@ -2,10 +2,10 @@ use super::AuthorisedAgentsPath;
 use crate::{
     AppError, AppStore, Context, HtmlTemplate,
     authorised_agents::AuthorisedAgent,
+    common::Problematic,
     filters,
     list_submitters::ListSubmitter,
     political_groups::{PoliticalGroup, PoliticalGroupSteps},
-    submit::Problematic,
 };
 use askama::Template;
 use axum::response::IntoResponse;
