+++ src/app/candidate_lists/pages/view.rs
@@ -4,8 +4,8 @@ use axum::response::IntoResponse;
 use crate::{
     AppError, Context, HtmlTemplate,
     candidate_lists::{CandidateList, FullCandidateList, pages::ViewCandidateListPath},
+    common::Problematic,
     filters,
-    submit::Problematic,
 };
 
 #[derive(Template)]
