+++ src/app/common/structs/mod.rs
@@ -15,6 +15,7 @@ mod last_name_prefix;
 mod name;
 mod postal_code;
 mod previous_election_results;
+mod problems;
 mod utc_date_time;
 
 pub use address::{Address, DutchAddress, InternationalAddress};
@@ -36,4 +37,5 @@ pub use last_name_prefix::LastNamePrefix;
 pub use name::FullName;
 pub use postal_code::{InternationalPostalCode, PostalCode};
 pub use previous_election_results::PreviousElectionResults;
+pub use problems::{PotentialProblems, Problematic, Severity};
 pub use utc_date_time::UtcDateTime;
