+++ src/app/submit/mod.rs
@@ -7,8 +7,5 @@ mod structs;
 pub use pages::{SubmitPath, router};
 pub use structs::{
     documents::DocumentData,
-    potential_problems::{
-        GeneralProblems, ListProblems, PersonProblems, PotentialProblems, Problematic, Problems,
-        Severity,
-    },
+    potential_problems::{GeneralProblems, ListProblems, PersonProblems, Problems},
 };
