+++ src/app/submit/structs/documents.rs
@@ -1,23 +1,20 @@
 use crate::{
     AppError, AppStore, Context, ElectionConfig, TypstRenderer,
     candidate_lists::{CandidateListId, FullCandidateList},
-    common::PreviousElectionResults,
+    common::{PreviousElectionResults, Problematic},
     core::{ModelLocale, Pdf, ZipResponseWriter},
-    submit::{
-        Problematic,
-        structs::{
-            eml210::Eml210,
-            h1::H1,
-            h3_1::H31,
-            h4::H4,
-            h9::H9,
-            typst_authorised_agent::TypstAuthorisedAgent,
-            typst_candidate::{TypstCandidate, ordered_candidates},
-            typst_datetime::TypstDatetime,
-            typst_detailed_candidate::TypstDetailedCandidate,
-            typst_electoral_districts::TypstElectoralDistricts,
-            typst_person::TypstPerson,
-        },
+    submit::structs::{
+        eml210::Eml210,
+        h1::H1,
+        h3_1::H31,
+        h4::H4,
+        h9::H9,
+        typst_authorised_agent::TypstAuthorisedAgent,
+        typst_candidate::{TypstCandidate, ordered_candidates},
+        typst_datetime::TypstDatetime,
+        typst_detailed_candidate::TypstDetailedCandidate,
+        typst_electoral_districts::TypstElectoralDistricts,
+        typst_person::TypstPerson,
     },
     utils::slugify_teletex,
 };
