+++ src/app/submit/structs/potential_problems.rs
@@ -1,15 +1,56 @@
 use axum_extra::routing::TypedPath as _;
 
 use crate::{
-    AppStore, ElectoralDistrict, Locale, QueryParamState,
+    AppStore, QueryParamState,
     authorised_agents::AuthorisedAgent,
     candidate_lists::{CandidateList, CandidateListSummary},
-    common::DateOfBirth,
+    common::{PotentialProblems, Problematic, Severity},
     list_submitters::ListSubmitter,
     persons::Person,
     political_groups::PoliticalGroup,
-    trans,
 };
+
+impl PotentialProblems {
+    pub fn candidate_list_fix_path(&self, list: &CandidateList) -> String {
+        match self {
+            PotentialProblems::NoCandidates => list.view_path().to_string(),
+            PotentialProblems::TooManyCandidates { actual, max } => {
+                let overflow = actual.saturating_sub(*max);
+                list.view_path()
+                    .with_query_params(QueryParamState::highlight_last(overflow))
+                    .to_string()
+            }
+            PotentialProblems::FewCandidatesWithFirstName { .. }
+            | PotentialProblems::FewCandidatesWithoutFirstName { .. }
+            | PotentialProblems::FewCandidatesWithGender { .. }
+            | PotentialProblems::FewCandidatesWithoutGender { .. } => list.view_path().to_string(),
+            _ => list.update_path().to_string(),
+        }
+    }
+
+    pub fn person_fix_path(&self, person: &Person) -> String {
+        match self {
+            PotentialProblems::NoStreetName(_)
+            | PotentialProblems::NoHouseNumber(_)
+            | PotentialProblems::NoPostalCode(_)
+            | PotentialProblems::NoLocality(_)
+            | PotentialProblems::NoCountry(_) => person.update_address_path().to_string(),
+            PotentialProblems::NoRepresentative | PotentialProblems::RepresentativeProblem(_) => {
+                person.update_representative_path().to_string()
+            }
+            _ => person.update_path().to_string(),
+        }
+    }
+
+    pub fn general_fix_path(&self) -> String {
+        match self {
+            PotentialProblems::NoAuthorisedAgent => AuthorisedAgent::list_path().to_string(),
+            PotentialProblems::NoListSubmitter => ListSubmitter::update_path().to_string(),
+            PotentialProblems::NoSubstituteSubmitter => ListSubmitter::view_path().to_string(),
+            _ => PoliticalGroup::update_path().to_string(),
+        }
+    }
+}
 
 /// Aggregation struct for everything that can be missing or incomplete for a list submission
 #[derive(Debug)]
@@ -89,7 +130,12 @@ impl Problems {
 
     pub fn models_downloadable(&self) -> bool {
         let candidate_iter = self.candidates.iter().flat_map(|ci| &ci.problems);
-        let list_iter = self.lists.iter().flat_map(|ci| &ci.problems);
+        // Lists without candidates cannot produce exports, so their errors don't block downloads
+        let list_iter = self
+            .lists
+            .iter()
+            .filter(|li| !li.problems.contains(&PotentialProblems::NoCandidates))
+            .flat_map(|ci| &ci.problems);
         let general_iter = self.general.flatten();
 
         !candidate_iter
@@ -123,6 +169,7 @@ impl GeneralProblems {
         result
     }
 }
+
 #[derive(Debug)]
 pub struct PersonProblems<T> {
     pub person: T,
@@ -142,310 +189,11 @@ pub struct ListProblems {
     pub problems: Vec<PotentialProblems>,
 }
 
-#[derive(Clone, PartialEq, Debug)]
-pub enum PotentialProblems {
-    // candidate list
-    NoCandidates,
-    TooManyCandidates { actual: usize, max: usize },
-    DuplicateDistricts { duplicates: Vec<ElectoralDistrict> },
-    NoDistricts,
-    FewCandidatesWithFirstName { count: usize, total: usize },
-    FewCandidatesWithoutFirstName { count: usize, total: usize },
-    FewCandidatesWithGender { count: usize, total: usize },
-    FewCandidatesWithoutGender { count: usize, total: usize },
-
-    // political group
-    NoLegalName,
-    NoDisplayName,
-    NoPreviousElectionResults,
-    NoAuthorisedAgent,
-    NoListSubmitter,
-    NoSubstituteSubmitter,
-
-    // representative wrapper
-    RepresentativeProblem(Box<PotentialProblems>),
-
-    // personal data
-    NoBsn,
-    VeryOldDateOfBirth,
-    NoPlaceOfResidence,
-    NoCountryOfResidence,
-    NoDateOfBirth,
-    NoRepresentative,
-
-    // name related
-    NoInitials(Severity),
-    NoLastName(Severity),
-
-    // address related
-    NoStreetName(Severity),
-    NoHouseNumber(Severity),
-    NoPostalCode(Severity),
-    NoLocality(Severity),
-    NoCountry(Severity),
-}
-
-impl PotentialProblems {
-    pub fn translate(&self, locale: &Locale) -> String {
-        match self {
-            // candidate list
-            PotentialProblems::NoCandidates => trans!("problems.no_candidates", *locale),
-            PotentialProblems::TooManyCandidates { actual, max } => {
-                trans!("problems.too_many_candidates", *locale, actual, max)
-            }
-            PotentialProblems::DuplicateDistricts { duplicates } => {
-                let districts = duplicates
-                    .iter()
-                    .map(|d| d.title((*locale).into()))
-                    .collect::<Vec<_>>()
-                    .join(", ");
-                trans!("problems.duplicate_districts", *locale, districts)
-            }
-            PotentialProblems::NoDistricts => trans!("problems.no_districts", *locale),
-            PotentialProblems::FewCandidatesWithFirstName { count, total } => {
-                if *count == 1 {
-                    trans!(
-                        "problems.few_candidates_with_first_name_one",
-                        *locale,
-                        total
-                    )
-                } else {
-                    trans!(
-                        "problems.few_candidates_with_first_name",
-                        *locale,
-                        count,
-                        total
-                    )
-                }
-            }
-            PotentialProblems::FewCandidatesWithoutFirstName { count, total } => {
-                if *count == 1 {
-                    trans!(
-                        "problems.few_candidates_without_first_name_one",
-                        *locale,
-                        total
-                    )
-                } else {
-                    trans!(
-                        "problems.few_candidates_without_first_name",
-                        *locale,
-                        count,
-                        total
-                    )
-                }
-            }
-            PotentialProblems::FewCandidatesWithGender { count, total } => {
-                if *count == 1 {
-                    trans!("problems.few_candidates_with_gender_one", *locale, total)
-                } else {
-                    trans!("problems.few_candidates_with_gender", *locale, count, total)
-                }
-            }
-            PotentialProblems::FewCandidatesWithoutGender { count, total } => {
-                if *count == 1 {
-                    trans!("problems.few_candidates_without_gender_one", *locale, total)
-                } else {
-                    trans!(
-                        "problems.few_candidates_without_gender",
-                        *locale,
-                        count,
-                        total
-                    )
-                }
-            }
-
-            // political group
-            PotentialProblems::NoLegalName => trans!("problems.no_legal_name", *locale),
-            PotentialProblems::NoDisplayName => trans!("problems.no_display_name", *locale),
-            PotentialProblems::NoPreviousElectionResults => {
-                trans!("problems.no_previous_election_results", *locale)
-            }
-            PotentialProblems::NoListSubmitter => trans!("problems.no_list_submitter", *locale),
-            PotentialProblems::NoAuthorisedAgent => trans!("problems.no_authorised_agent", *locale),
-            PotentialProblems::NoSubstituteSubmitter => {
-                trans!("problems.no_substitute_submitter", *locale)
-            }
-
-            // representative wrapper
-            PotentialProblems::RepresentativeProblem(inner) => {
-                let label = trans!("problems.representative", *locale);
-                let problem = inner.translate(locale);
-                format!("{label}: {problem}")
-            }
-
-            // personal data
-            PotentialProblems::NoBsn => trans!("problems.no_bsn", *locale),
-            PotentialProblems::VeryOldDateOfBirth => {
-                trans!(
-                    "problems.very_old_date_of_birth",
-                    *locale,
-                    DateOfBirth::WARN_AGE
-                )
-            }
-            PotentialProblems::NoPlaceOfResidence => {
-                trans!("problems.no_place_of_residence", *locale)
-            }
-            PotentialProblems::NoCountryOfResidence => {
-                trans!("problems.no_country_of_residence", *locale)
-            }
-            PotentialProblems::NoDateOfBirth => trans!("problems.no_date_of_birth", *locale),
-            PotentialProblems::NoRepresentative => trans!("problems.no_representative", *locale),
-
-            // name related
-            PotentialProblems::NoInitials(_) => trans!("problems.no_initials", *locale),
-            PotentialProblems::NoLastName(_) => trans!("problems.no_last_name", *locale),
-
-            // address related
-            PotentialProblems::NoStreetName(_) => trans!("problems.no_street_name", *locale),
-            PotentialProblems::NoHouseNumber(_) => trans!("problems.no_house_number", *locale),
-            PotentialProblems::NoPostalCode(_) => trans!("problems.no_postal_code", *locale),
-            PotentialProblems::NoLocality(_) => trans!("problems.no_locality", *locale),
-            PotentialProblems::NoCountry(_) => trans!("problems.no_country", *locale),
-        }
-    }
-
-    pub fn candidate_list_fix_path(&self, list: &CandidateList) -> String {
-        match self {
-            PotentialProblems::NoCandidates => list.view_path().to_string(),
-            PotentialProblems::TooManyCandidates { actual, max } => {
-                let overflow = actual.saturating_sub(*max);
-                list.view_path()
-                    .with_query_params(QueryParamState::highlight_last(overflow))
-                    .to_string()
-            }
-            PotentialProblems::FewCandidatesWithFirstName { .. }
-            | PotentialProblems::FewCandidatesWithoutFirstName { .. }
-            | PotentialProblems::FewCandidatesWithGender { .. }
-            | PotentialProblems::FewCandidatesWithoutGender { .. } => list.view_path().to_string(),
-            _ => list.update_path().to_string(),
-        }
-    }
-
-    pub fn person_fix_path(&self, person: &Person) -> String {
-        match self {
-            PotentialProblems::NoStreetName(_)
-            | PotentialProblems::NoHouseNumber(_)
-            | PotentialProblems::NoPostalCode(_)
-            | PotentialProblems::NoLocality(_)
-            | PotentialProblems::NoCountry(_) => person.update_address_path().to_string(),
-            PotentialProblems::NoRepresentative | PotentialProblems::RepresentativeProblem(_) => {
-                person.update_representative_path().to_string()
-            }
-            _ => person.update_path().to_string(),
-        }
-    }
-
-    pub fn general_fix_path(&self) -> String {
-        match self {
-            PotentialProblems::NoAuthorisedAgent => AuthorisedAgent::list_path().to_string(),
-            PotentialProblems::NoListSubmitter => ListSubmitter::update_path().to_string(),
-            PotentialProblems::NoSubstituteSubmitter => ListSubmitter::view_path().to_string(),
-            _ => PoliticalGroup::update_path().to_string(),
-        }
-    }
-
-    pub fn severity_class(&self) -> &'static str {
-        match self.severity() {
-            Severity::Info => "info",
-            Severity::Warn => "warning",
-            Severity::Error => "error",
-        }
-    }
-
-    pub fn severity(&self) -> Severity {
-        match &self {
-            // candidate list
-            PotentialProblems::NoCandidates => Severity::Error,
-            PotentialProblems::TooManyCandidates { .. } => Severity::Warn,
-            PotentialProblems::DuplicateDistricts { .. } => Severity::Error,
-            PotentialProblems::NoDistricts => Severity::Error,
-            PotentialProblems::FewCandidatesWithFirstName { .. } => Severity::Info,
-            PotentialProblems::FewCandidatesWithoutFirstName { .. } => Severity::Info,
-            PotentialProblems::FewCandidatesWithGender { .. } => Severity::Info,
-            PotentialProblems::FewCandidatesWithoutGender { .. } => Severity::Info,
-
-            // political group
-            PotentialProblems::NoLegalName => Severity::Warn,
-            PotentialProblems::NoDisplayName => Severity::Error,
-            PotentialProblems::NoPreviousElectionResults => Severity::Info,
-            PotentialProblems::NoListSubmitter => Severity::Error,
-            PotentialProblems::NoAuthorisedAgent => Severity::Warn,
-            PotentialProblems::NoSubstituteSubmitter => Severity::Info,
-
-            // representative wrapper
-            PotentialProblems::RepresentativeProblem(inner) => inner.severity(),
-
-            // personal data
-            PotentialProblems::NoBsn => Severity::Warn,
-            PotentialProblems::VeryOldDateOfBirth => Severity::Warn,
-            PotentialProblems::NoPlaceOfResidence => Severity::Error,
-            PotentialProblems::NoCountryOfResidence => Severity::Error,
-            PotentialProblems::NoDateOfBirth => Severity::Error,
-            PotentialProblems::NoRepresentative => Severity::Warn,
-
-            // name related
-            PotentialProblems::NoInitials(severity) => *severity,
-            PotentialProblems::NoLastName(severity) => *severity,
-
-            // address related
-            PotentialProblems::NoStreetName(severity) => *severity,
-            PotentialProblems::NoHouseNumber(severity) => *severity,
-            PotentialProblems::NoPostalCode(severity) => *severity,
-            PotentialProblems::NoLocality(severity) => *severity,
-            PotentialProblems::NoCountry(severity) => *severity,
-        }
-    }
-}
-
-/// Problem severities, in increasing order of severity
-#[derive(PartialEq, Eq, PartialOrd, Ord, Clone, Copy, Debug)]
-pub enum Severity {
-    Info,
-    Warn,
-    Error,
-}
-
-impl Severity {
-    pub fn class(&self) -> &'static str {
-        match self {
-            Severity::Info => "info",
-            Severity::Warn => "warning",
-            Severity::Error => "error",
-        }
-    }
-}
-
-pub trait Problematic {
-    /// Returns all potential problems of its own and of all children
-    fn get_problems(&self) -> Vec<PotentialProblems>;
-
-    /// Returns true if there are no problems
-    fn is_all_good(&self) -> bool {
-        self.get_problems().is_empty()
-    }
-
-    /// Returns the highest severity of the problems, or None if there are no problems
-    fn highest_severity(&self) -> Option<Severity> {
-        self.get_problems().into_iter().map(|p| p.severity()).max()
-    }
-
-    /// Returns the CSS class associated with the highest severity
-    fn highest_severity_class(&self) -> &'static str {
-        match self.highest_severity() {
-            None => "ok",
-            Some(severity) => severity.class(),
-        }
-    }
-
-    fn has_severity_or_higher(&self, severity: Severity) -> bool {
-        self.get_problems().iter().any(|p| p.severity() >= severity)
-    }
-}
-
 #[cfg(test)]
 mod tests {
     use crate::{
         candidate_lists::CandidateListId,
+        common::Severity,
         persons::PersonId,
         test_utils::{sample_candidate_list, sample_person},
     };
